refactor(hub): policy engine to with-party/policy — kernel is spine + enclave
