+++ packages/hub/__tests__/auth-introspection.test.ts
@@ -16,7 +16,7 @@ import {
 import {
   saveVaultPolicy,
   PERSONAL_POLICY,
-} from '../src/kernel/policy/index.js'
+} from '../src/with-party/policy/index.js'
 
 function inlineMemory(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
