+++ packages/hub/__tests__/liberate.test.ts
@@ -30,7 +30,7 @@
 import { describe, it, expect, beforeEach } from 'vitest'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/kernel/types.js'
 import { ConflictError, PermissionDeniedError } from '../src/kernel/errors.js'
-import { PolicyDeniedError } from '../src/kernel/policy/errors.js'
+import { PolicyDeniedError } from '../src/kernel/errors.js'
 import { createNoydb } from '../src/kernel/noydb.js'
 import type { Noydb } from '../src/kernel/noydb.js'
 import { withHistory } from '../src/with-commit/history/index.js'
