+++ packages/hub/__tests__/strict-policy-grant-revoke.test.ts
@@ -10,8 +10,8 @@
 import { describe, it, expect } from 'vitest'
 import type { NoydbStore, EncryptedEnvelope } from '../src/kernel/types.js'
 import { createNoydb } from '../src/kernel/noydb.js'
-import { PolicyDeniedError } from '../src/kernel/policy/errors.js'
-import { STRICT_POLICY } from '../src/kernel/policy/presets.js'
+import { PolicyDeniedError } from '../src/kernel/errors.js'
+import { STRICT_POLICY } from '../src/with-party/policy/presets.js'
 
 function inlineMemory(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
