+++ packages/hub/__tests__/withdraw-accessible.test.ts
@@ -6,10 +6,10 @@
 import { describe, it, expect } from 'vitest'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/kernel/types.js'
 import { ConflictError } from '../src/kernel/errors.js'
-import { PolicyDeniedError } from '../src/kernel/policy/errors.js'
+import { PolicyDeniedError } from '../src/kernel/errors.js'
 import { createNoydb } from '../src/kernel/noydb.js'
 import { withPortability } from '../src/with-audit/portability/index.js'
-import type { VaultPolicy } from '../src/kernel/policy/types.js'
+import type { VaultPolicy } from '../src/kernel/types.js'
 import { readNoydbBundle } from '../src/with-pod/bundle.js'
 
 function makeStore(): NoydbStore {
