--- packages/hub/src/kernel/policy/types.ts
@@ -1,230 +1,0 @@
-/**
- * Policy gate DSL types.
- *
- * Sensitive operations (rotate the passphrase, enroll an authenticator,
- * export plaintext, grant a user, …) are gated by a typed policy
- * object. The developer supplies a {@link VaultPolicy} at vault
- * creation; the hub merges it onto a built-in preset and persists the
- * merged document at `_meta/policy`.
- *
- * @see docs/services/session-tiers.md → Policy gates DSL
- *
- * @module
- */
-import type { PassphrasePolicy } from '../validation.js'
-
-/**
- * A single factor surface — the proof an actor presents at gate time.
- *
- * | Kind | Source | Off-device? |
- * |---|---|---|
- * | `totp` | RFC 6238 authenticator app (Google Auth, 1Password) | yes |
- * | `email-otp` | one-time code mailed to the user | yes |
- * | `recovery` | printable Base32 code (`@noy-db/on-recovery`) | yes (paper) |
- * | `shamir` | k-of-n threshold share (`@noy-db/on-shamir`) | yes |
- * | `webauthn-roaming` | hardware key (YubiKey, SoloKey, Titan) | yes (key portable) |
- * | `webauthn-platform` | platform passkey (Touch ID, Face ID, Hello) | no (device-bound) |
- * | `password` | tier-2 password (`@noy-db/on-password`) | no |
- * | `pin` | tier-3 quick-resume PIN (`@noy-db/on-pin`) | no |
- *
- * Off-device kinds (TOTP, email-OTP, recovery, shamir, roaming WebAuthn)
- * are the strongest factor proofs because they require something
- * separate from the device the user just unlocked. Platform / password /
- * PIN are useful for "fresh proof of *this* user" but don't bind across
- * devices — policies can require ANY of them or insist on a count of 2
- * to force a mix.
- *
- * `webauthn-platform`, `password`, `pin` — for consumers with no
- * off-device infrastructure (no TOTP, no email-OTP, paper recovery not
- * enrolled) who want to require "any second factor I have wired"
- * without losing the freshness guarantee.
- */
-export type FactorKind =
-  | 'totp'
-  | 'email-otp'
-  | 'recovery'
-  | 'shamir'
-  | 'webauthn-roaming'
-  | 'webauthn-platform'
-  | 'password'
-  | 'pin'
-
-/**
- * One factor requirement entry. The default is "any one of the listed
- * factors, fresh within the last 5 minutes". Bumping `count` requires N
- * distinct fresh proofs; bumping `freshnessMs` widens the acceptance
- * window.
- */
-export interface FactorRequirement {
-  readonly anyOf: ReadonlyArray<FactorKind>
-  /** Number of distinct factors required. Default 1. */
-  readonly count?: number
-  /** How recent each proof must be. Default 5 minutes. */
-  readonly freshnessMs?: number
-}
-
-/** Soft signals layered on top of the gate verdict — never block on their own. */
-export interface WarningRules {
-  /** Behavior on shared-device tier-1 ops. `'block'` raises a `PolicyDeniedError`. */
-  readonly sharedDevice?: 'warn' | 'block'
-  /** Behavior on weak tier-2 (e.g. password-only) for sensitive ops. */
-  readonly weakAuthenticator?: 'warn' | 'block'
-}
-
-/**
- * Policy applied to one named gate. `enabled: false` disables the
- * action entirely (useful in managed-passphrase mode where rotation is
- * impossible by construction).
- */
-export interface GatePolicy {
-  /** Minimum tier the active session must hold. */
-  readonly minTier: 1 | 2 | 3
-  /** Extra freshness-bound proofs required at gate time. */
-  readonly factors?: ReadonlyArray<FactorRequirement>
-  readonly warn?: WarningRules
-  readonly enabled?: boolean
-}
-
-/**
- * Built-in gate names. App-defined gates live in the `app:*` namespace
- * and use the same engine; the engine treats unknown names with no
- * configured policy as "no gate" (no-op).
- */
-export type BuiltInGateName =
-  | 'rotate-passphrase'
-  | 'recover-passphrase'
-  | 'enroll-authenticator'
-  | 'remove-authenticator'
-  /**
-   * Authorize a deliberate paper-recovery-code regeneration —
-   * `db.rotateRecovery`. Symmetric to `rotate-passphrase` for
-   * the case where the user remembers their passphrase but wants a
-   * fresh sheet (lost the printout, suspect compromise of the off-site
-   * copy). PERSONAL allows tier-1; STRICT requires an off-device
-   * factor so a stolen unlocked laptop cannot silently mint a new
-   * sheet for an attacker.
-   */
-  | 'rotate-recovery'
-  /**
-   * Authorize a meta-only mutation on an existing authenticator slot —
-   * `db.updateAuthenticator`. The slot's wrap material, id, and
-   * method are immutable through this gate; only the `meta` blob
-   * (nicknames, method-specific labels) can change. Anti-slot-swap
-   * guard is preserved structurally regardless of this gate's
-   * settings.
-   */
-  | 'update-authenticator'
-  | 'rotate-unlock'
-  | 'enroll-user'
-  | 'revoke-user'
-  | 'export-bundle'
-  | 'export-plaintext'
-  | 'view-user-auth'
-  /** Authorize a write to one's own user envelope. */
-  | 'edit-own-profile'
-  /** Authorize reading other principals' user envelopes. */
-  | 'view-team-profiles'
-  /**
-   * Authorize an atomic peer-recovery — `db.recoverUser`.
-   * Distinct from `revoke-user` because peer-recovery is intentional
-   * re-issuance of someone's keyring under a temp passphrase, NOT
-   * removal. Allows owner→owner natively (matches the threat model:
-   * a co-owner explicitly recovering another co-owner). Ships with a
-   * factor-proof default in `STRICT_POLICY` so the issuer must
-   * affirmatively prove identity at the moment of recovery.
-   */
-  | 'peer-recover-user'
-  /**
-   * Authorize a post-grant identity mutation — `db.updateUser`.
-   * Covers `role`, `displayName`, `permissions` changes on an existing
-   * keyring. Pure plaintext-header rewrite — no DEKs touched, no KEK
-   * required. The role-elevation guard inside the implementation
-   * mirrors `db.grant`'s hierarchy (admin cannot promote to owner)
-   * regardless of this gate's settings.
-   */
-  | 'update-user'
-  /**
-   * Authorize a non-owner's self-service **destructive** withdrawal —
-   * `vault.user.unilateralWithdrawal`. The actor exports their
-   * own re-keyed copy and then removes (delete-closure) or freezes the
-   * source records. Because it both egresses data AND destroys the
-   * firm's live copy, it MUST fail closed: undefined in a policy = denied.
-   * Hosts opt in explicitly (and typically pin `minTier`/factor proofs).
-   */
-  | 'client-unilateral-withdraw'
-  /**
-   * Authorize FILING a two-party withdrawal request —
-   * `vault.user.requestWithdrawal`. Non-destructive (writes a
-   * pending request only); enabled by default so a read-only client can ask.
-   */
-  | 'user-request-withdrawal'
-  /**
-   * Authorize DECIDING a two-party withdrawal request (approve/reject) —
-   * `vault.user.approveWithdrawal` / `rejectWithdrawal`. The approve
-   * path is destructive (extract-and-dispose under firm authority), so it
-   * defaults to a tier-2 floor; owner/admin role is enforced structurally.
-   */
-  | 'approve-user-withdrawal'
-  /**
-   * Authorize minting a **custodian** — `db.grantCustodian` (FR-6). The
-   * custodian is the de-facto operational authority on a sealed-owner (Deed)
-   * vault, so granting one is an ownership-level act: this gate MUST fail
-   * closed (undefined in a policy = denied) and owner-only role is enforced
-   * structurally. Hosts opt in explicitly, typically pinning factor proofs.
-   */
-  | 'grant-custodian'
-  /**
-   * Authorize the audited **Liberate** ceremony — `vault.custody.liberate`
-   * (FR-6). The custodian (holding the live DEKs) claims ownership of a
-   * sealed-owner vault under a recorded legal basis, minting a NEW owner
-   * keyring. Destructive-of-the-old-ownership and irreversible, so it MUST
-   * fail closed (undefined = denied); the caller-is-custodian check is
-   * enforced structurally in the ceremony.
-   */
-  | 'liberate-vault'
-
-/** Either a built-in gate name or an `app:*` custom gate. */
-export type GateName = BuiltInGateName | `app:${string}`
-
-/**
- * Top-level policy object. Persisted at `_meta/policy` once at vault
- * creation. The `passphrase` block configures the strength rules
- * applied at every passphrase ingress; `gates` configures
- * the action-level requirements.
- */
-export interface VaultPolicy {
-  readonly passphrase?: PassphrasePolicy
-  readonly gates: Partial<Record<GateName, GatePolicy>>
-}
-
-/** Concrete proof an actor presents to {@link checkGate}. */
-export interface FactorProof {
-  readonly kind: FactorKind
-  /** ISO-8601 timestamp the proof was minted at. Compared against `freshnessMs`. */
-  readonly mintedAt?: string
-  /** Method-specific payload. The engine treats it as opaque — verification is delegated. */
-  readonly payload?: unknown
-}
-
-/**
- * Bundle of factor proofs + session-context flags passed to a gated
- * Noydb method. Used as the optional last parameter of every method
- * that runs through `checkGate`: `db.grant`, `db.revoke`, `db.updateUser`,
- * `db.enrollAuthenticator`, `db.removeAuthenticator`, `db.updateAuthenticator`,
- * `db.enrollWebAuthn`, `db.rotatePassphrase`, `db.recoverPassphrase`,
- * `db.recoverUser`, `db.enrollUnlock`, `db.describeUserAuth`,
- * `db.describeAllUsersAuth`.
- *
- * Previously this type was inlined at every call site as
- * `{ factors?: ReadonlyArray<FactorProof>; sharedDevice?: boolean }`
- * and parameter names alternated between `factors` and `presented`.
- * Now exported so consumers can name their helpers and so the param
- * name converges to `factors` everywhere.
- */
-export interface FactorProofBundle {
-  readonly factors?: ReadonlyArray<FactorProof>
-  readonly sharedDevice?: boolean
-}
-
-/** Active session tier — what the engine compares against `gate.minTier`. */
-export type ActiveTier = 1 | 2 | 3
